Add test for cookies landing page
