+++ packages/extension/src/view/devtools/components/tabContent/cookies/cookiesLanding/index.tsx
@@ -38,7 +38,7 @@ const CookiesLanding = () => {
   const cookiesStatsComponents = prepareCookieStatsComponents(cookieStats);
 
   return (
-    <div className="h-full w-full">
+    <div className="h-full w-full" data-testid="cookies-landing">
       <LandingHeader
         cookieStats={cookieStats}
         cookiesStatsComponents={cookiesStatsComponents}
