+++ packages/extension/src/view/devtools/components/tabContent/cookies/cookiesLanding/landingHeader/index.tsx
@@ -62,7 +62,10 @@ const LandingHeader = ({
   ];
 
   return (
-    <div className="flex justify-center border-b border-hex-gray pt-5 pb-5">
+    <div
+      className="flex justify-center border-b border-hex-gray pt-5 pb-5"
+      data-testid="cookies-landing-header"
+    >
       <div className="lg:max-w-[729px] flex gap-9 px-4">
         {dataMapping.map((circleData, index) => {
           return (
