+++ packages/extension/src/view/devtools/components/tabContent/cookies/tests/cookieTab.tsx
@@ -19,164 +19,20 @@
 import React from 'react';
 import { fireEvent, render, screen, within } from '@testing-library/react';
 import '@testing-library/jest-dom';
-import { type Cookie as ParsedCookie } from 'simple-cookie';
 import SinonChrome from 'sinon-chrome';
 
 /**
  * Internal dependencies.
  */
 import CookieTab from '..';
-import type { CookieStoreContext } from '../../../../stateProviders/syncCookieStore';
-import { emptyAnalytics } from '../../../../../../worker/findAnalyticsMatch';
 import CookieDetails from '../cookiesListing/cookieDetails';
 import { useContentPanelStore } from '../../../../stateProviders/contentPanelStore';
 import Details from '../cookiesListing/cookieDetails/details';
-
-const emptyCookie = {
-  name: '',
-  value: '',
-  domain: '',
-  samesite: '',
-  secure: false,
-  httponly: false,
-  path: '',
-  expires: '',
-};
-
-const uncategorized1pCookie: ParsedCookie = {
-  ...emptyCookie,
-  name: '_cb',
-  value: 'v1%3A168740954476563235',
-  domain: '.cnn.com',
-};
-
-const uncategorized3pCookie: ParsedCookie = {
-  ...emptyCookie,
-  name: 'pubsyncexp',
-  value: 'uncategorized3pCookie',
-  domain: '.ads.pubmatic.com',
-};
-
-const known1pCookie: ParsedCookie = {
-  ...emptyCookie,
-  name: '__qca',
-  value: 'known1pCookie',
-  domain: '.cnn.com',
-};
-
-const known3pCookie: ParsedCookie = {
-  ...emptyCookie,
-  name: 'KRTBCOOKIE_290',
-  value: 'known3pCookie',
-  domain: '.pubmatic.com',
-};
-
-const known3pCookieWithValue: ParsedCookie = {
-  ...emptyCookie,
-  name: 'KRTBCOOKIE_290',
-  value: 'known3p_Cookie-with%20value',
-  domain: '.pubmatic.com',
-};
-
-const mockResponse: {
-  tabCookies: NonNullable<CookieStoreContext['state']['tabCookies']>;
-  tabUrl: NonNullable<CookieStoreContext['state']['tabUrl']>;
-  tabFrames: NonNullable<CookieStoreContext['state']['tabFrames']>;
-  selectedFrame: NonNullable<CookieStoreContext['state']['selectedFrame']>;
-} = {
-  tabCookies: {
-    [uncategorized1pCookie.name]: {
-      parsedCookie: uncategorized1pCookie,
-      analytics: { ...emptyAnalytics },
-      url: 'https://edition.cnn.com/whatever/api',
-      headerType: 'response',
-      isFirstParty: true,
-      isIbcCompliant: true,
-      isCookieSet: true,
-      frameIdList: [1],
-    },
-    [uncategorized3pCookie.name]: {
-      parsedCookie: uncategorized3pCookie,
-      analytics: { ...emptyAnalytics },
-      url: 'https://api.pubmatic.com/whatever/api',
-      headerType: 'response',
-      isFirstParty: false,
-      isIbcCompliant: true,
-      isCookieSet: true,
-      frameIdList: [1],
-    },
-    [known1pCookie.name]: {
-      parsedCookie: known1pCookie,
-      analytics: {
-        platform: 'Quantcast',
-        category: 'Marketing',
-        name: '__qca',
-        domain: "Advertiser's website domain",
-        description:
-          'This cookie is set by Quantcast, who present targeted advertising. Stores browser and HTTP request information.',
-        retention: '1 year',
-        dataController: 'Quantcast',
-        gdprUrl: 'https://www.quantcast.com/privacy/',
-        wildcard: '0',
-      },
-      url: 'https://edition.cnn.com/whatever/api',
-      headerType: 'response',
-      isFirstParty: true,
-      isIbcCompliant: true,
-      isCookieSet: true,
-      frameIdList: [1],
-    },
-    [known3pCookie.name]: {
-      parsedCookie: known3pCookie,
-      analytics: {
-        platform: 'PubMatic',
-        category: 'Marketing',
-        name: 'KRTBCOOKIE_*',
-        domain: 'pubmatic.com',
-        description:
-          "Registers a unique ID that identifies the user's device during return visits across websites that use the same ad network. The ID is used to allow targeted ads.",
-        retention: '29 days',
-        dataController: 'Pubmatic',
-        gdprUrl: 'N/A',
-        wildcard: '1',
-      },
-      url: 'https://api.pubmatic.com/whatever/api',
-      headerType: 'response',
-      isFirstParty: false,
-      isIbcCompliant: true,
-      isCookieSet: true,
-      frameIdList: [1],
-    },
-    [known3pCookieWithValue.name]: {
-      parsedCookie: known3pCookieWithValue,
-      analytics: {
-        platform: 'PubMatic',
-        category: 'Marketing',
-        name: 'KRTBCOOKIE_*',
-        domain: 'pubmatic.com',
-        description:
-          "Registers a unique ID that identifies the user's device during return visits across websites that use the same ad network. The ID is used to allow targeted ads.",
-        retention: '29 days',
-        dataController: 'Pubmatic',
-        gdprUrl: 'N/A',
-        wildcard: '1',
-      },
-      url: 'https://api.pubmatic.com/whatever/api',
-      headerType: 'response',
-      isFirstParty: false,
-      isIbcCompliant: true,
-      isCookieSet: true,
-      frameIdList: [1],
-    },
-  },
-  tabUrl: 'https://edition.cnn.com/',
-  tabFrames: {
-    'https://edition.cnn.com/': {
-      frameIds: [1],
-    },
-  },
-  selectedFrame: 'https://edition.cnn.com/',
-};
+import mockResponse, {
+  uncategorized1pCookie,
+  known1pCookie,
+  known3pCookieWithValue,
+} from '../../../../../../utils/test-data/cookieMockData';
 
 jest.mock('../../../../stateProviders/syncCookieStore', () => {
   return {
